refactor(cli): moved TypedTransactionArgs and related trait impl to common mod
